[MLIR][RemoveDeadValues] Add a prior pass PrivatizePublicFunction
A new pass is added before ReadDeadValues. It attempts to privatize the public function if
1) liveness is inter-procedural.
2) There is NonLive parameter.
3) There is at least one CallOp.

Our strategy is to separate interface and implementation.
A new private function is created and it takes the function body of the original one.
The original function invokes the private function and return whatever it returns.